refactor: VerifyStatus 위치 변경
- application.domain 하위 -> common 하위
