+++ src/main/java/com/example/solidconnection/application/repository/ApplicationRepository.java
@@ -1,7 +1,7 @@
 package com.example.solidconnection.application.repository;
 
 import com.example.solidconnection.application.domain.Application;
-import com.example.solidconnection.application.domain.VerifyStatus;
+import com.example.solidconnection.common.VerifyStatus;
 import com.example.solidconnection.common.exception.CustomException;
 import com.example.solidconnection.siteuser.domain.SiteUser;
 import org.springframework.data.jpa.repository.JpaRepository;
