+++ src/main/java/com/example/solidconnection/common/VerifyStatus.java
@@ -1,4 +1,4 @@
-package com.example.solidconnection.application.domain;
+package com.example.solidconnection.common;
 
 public enum VerifyStatus {
     PENDING, REJECTED, APPROVED
