+++ src/main/java/com/example/solidconnection/score/dto/GpaScoreStatusResponse.java
@@ -1,6 +1,6 @@
 package com.example.solidconnection.score.dto;
 
-import com.example.solidconnection.application.domain.VerifyStatus;
+import com.example.solidconnection.common.VerifyStatus;
 import com.example.solidconnection.score.domain.GpaScore;
 
 public record GpaScoreStatusResponse(
