+++ src/test/java/com/example/solidconnection/admin/dto/validation/RejectedReasonValidatorTest.java
@@ -2,7 +2,7 @@ package com.example.solidconnection.admin.dto.validation;
 
 import com.example.solidconnection.admin.dto.GpaScoreUpdateRequest;
 import com.example.solidconnection.admin.dto.LanguageTestScoreUpdateRequest;
-import com.example.solidconnection.application.domain.VerifyStatus;
+import com.example.solidconnection.common.VerifyStatus;
 import com.example.solidconnection.university.domain.LanguageTestType;
 import jakarta.validation.ConstraintViolation;
 import jakarta.validation.Validation;
