+++ src/test/java/com/example/solidconnection/score/fixture/GpaScoreFixture.java
@@ -1,7 +1,7 @@
 package com.example.solidconnection.score.fixture;
 
 import com.example.solidconnection.application.domain.Gpa;
-import com.example.solidconnection.application.domain.VerifyStatus;
+import com.example.solidconnection.common.VerifyStatus;
 import com.example.solidconnection.score.domain.GpaScore;
 import com.example.solidconnection.siteuser.domain.SiteUser;
 import lombok.RequiredArgsConstructor;
