+++ src/test/java/com/example/solidconnection/score/fixture/LanguageTestScoreFixture.java
@@ -1,10 +1,9 @@
 package com.example.solidconnection.score.fixture;
 
 import com.example.solidconnection.application.domain.LanguageTest;
-import com.example.solidconnection.application.domain.VerifyStatus;
+import com.example.solidconnection.common.VerifyStatus;
 import com.example.solidconnection.score.domain.LanguageTestScore;
 import com.example.solidconnection.siteuser.domain.SiteUser;
-import com.example.solidconnection.university.domain.LanguageTestType;
 import lombok.RequiredArgsConstructor;
 import org.springframework.boot.test.context.TestComponent;
