+++ src/test/java/com/example/solidconnection/score/fixture/LanguageTestScoreFixtureBuilder.java
@@ -1,7 +1,7 @@
 package com.example.solidconnection.score.fixture;
 
 import com.example.solidconnection.application.domain.LanguageTest;
-import com.example.solidconnection.application.domain.VerifyStatus;
+import com.example.solidconnection.common.VerifyStatus;
 import com.example.solidconnection.score.domain.LanguageTestScore;
 import com.example.solidconnection.score.repository.LanguageTestScoreRepository;
 import com.example.solidconnection.siteuser.domain.SiteUser;
