+++ src/test/java/com/example/solidconnection/score/service/ScoreServiceTest.java
@@ -1,6 +1,6 @@
 package com.example.solidconnection.score.service;
 
-import com.example.solidconnection.application.domain.VerifyStatus;
+import com.example.solidconnection.common.VerifyStatus;
 import com.example.solidconnection.s3.domain.ImgType;
 import com.example.solidconnection.s3.dto.UploadedFileUrlResponse;
 import com.example.solidconnection.s3.service.S3Service;
